moved Map class to runtime and moved tests to main in runtime everthing seems to be working as expected
