+++ packages/core/src/index.ts
@@ -14,4 +14,3 @@ import "./rotaryencoder"
 import "./button"
 import "./magneticfieldlevel"
 import "./buzzer"
-import "./map"
