--- packages/core/src/map.ts
@@ -1,51 +1,0 @@
-export class Map<K, V> {
-    private keys: K[] = []
-    private values: V[] = []
-
-    constructor(entries?: Array<readonly [K, V]> | null) {
-        if (entries) {
-            for (let i = 0; i < entries.length; i++) {
-                const [key, value] = entries[i]
-                this.set(key, value)
-            }
-        }
-    }
-
-    size(): number {
-        return this.keys.length
-    }
-
-    set(key: K, value: V): void {
-        const index = this.keys.indexOf(key)
-        if (index !== -1) {
-            this.values[index] = value
-        } else {
-            this.keys.push(key)
-            this.values.push(value)
-        }
-    }
-
-    get(key: K): V | undefined {
-        const index = this.keys.indexOf(key)
-        return index !== -1 ? this.values[index] : undefined
-    }
-
-    has(key: K): boolean {
-        return this.keys.indexOf(key) !== -1
-    }
-
-    delete(key: K): boolean {
-        const index = this.keys.indexOf(key)
-        if (index !== -1) {
-            this.keys.insert(index, -1)
-            this.values.insert(index, -1)
-            return true
-        }
-        return false
-    }
-
-    clear(): void {
-        this.keys = []
-        this.values = []
-    }
-}
